Move feature spec support files under system
We don't have feature specs anymore. This change also groups
multiple support helper files into a single one because they
are all related to checkout.
