+++ templates/spec/support/solidus_starter_frontend/system/assets.rb
@@ -1,8 +1,4 @@
 RSpec.configure do |config|
-  config.when_first_matching_example_defined(type: :feature) do
-    config.before(:suite) { Rails.application.precompiled_assets }
-  end
-
   config.when_first_matching_example_defined(type: :system) do
     config.before(:suite) { Rails.application.precompiled_assets }
   end
