+++ templates/spec/system/address_spec.rb
@@ -3,7 +3,7 @@
 require 'solidus_starter_frontend_spec_helper'
 
 RSpec.describe 'Address', type: :system, inaccessible: true do
-  include  SolidusStarterFrontend::SystemHelpers
+  include  SolidusStarterFrontend::System::CheckoutHelpers
 
   let!(:product) { create(:product, name: "RoR Mug") }
   let!(:order) { create(:order_with_totals, state: 'cart') }
